pyproject.toml
[project]
name = "transfer-files"
version = "0.1.0"
authors = [{ name = "José Roberto Ferreira" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paramiko>=4.0.0",
    "pymssql>=2.3.7",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = ["sqlacodegen>=3.0.0"]

[tool.ruff]
line-length = 120
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'
